refaktorisering + justeringar

+++ src/map/PathfinderWeighter.java
@@ -1,5 +1,5 @@
 package map;
 
 public interface PathfinderWeighter {
-  int weight(Region a, Region b);
+  float weight(Region a, Region b);
 }
